+++ .gitignore
@@ -1,0 +1,2 @@
+# Local Netlify folder
+.netlify
